no need for separate htsget_methods file
