+++ htsget_ingest.py
@@ -4,8 +4,129 @@ import auth
 import os
 import re
 import json
-from htsget_methods import post_object
 from ingest_result import IngestPermissionsException, IngestServerException, IngestUserException, IngestResult
+import requests
+from urllib.parse import urlparse
+
+
+CANDIG_URL = os.getenv("CANDIG_URL", "")
+HTSGET_URL = CANDIG_URL + "/genomics"
+HOSTNAME = HTSGET_URL.replace(f"{urlparse(CANDIG_URL).scheme}://","")
+
+def post_object(token, genomic_sample, clinical_samples, dataset, ref_genome="hg38", force=False):
+    headers = {"Authorization": f"Bearer {token}"}
+    print(f"working on {genomic_sample['id']}")
+    url = f"{HTSGET_URL}/ga4gh/drs/v1/objects"
+
+    # file object:
+    access_method = {}
+    if genomic_sample['file_access'].startswith("file://"):
+        access_method["access_url"] = {
+            "headers": [],
+            "url": genomic_sample['file_access']
+        }
+        access_method["type"] = "file"
+
+    else:
+        access_method["access_id"] = genomic_sample['file_access']
+        access_method["type"] = "s3"
+
+    obj = {
+        "access_methods": [
+            access_method
+        ],
+        "id": genomic_sample["id"],
+        "name": genomic_sample["id"],
+        "description": "variant",
+        "cohort": dataset,
+        "version": "v1"
+    }
+    response = requests.post(url, json=obj, headers=headers)
+    if response.status_code > 200:
+        return response
+
+    # index object:
+    access_method = {}
+    if genomic_sample['index_access'].startswith("file://"):
+        access_method["access_url"] = {
+            "headers": [],
+            "url": genomic_sample['index_access']
+        }
+        access_method["type"] = "file"
+
+    else:
+        access_method["access_id"] = genomic_sample['index_access']
+        access_method["type"] = "s3"
+
+    obj = {
+        "access_methods": [
+            access_method
+        ],
+        "id": genomic_sample["index"],
+        "name": genomic_sample["index"],
+        "description": "index",
+        "cohort": dataset,
+        "version": "v1"
+    }
+
+    response = requests.post(url, json=obj, headers=headers)
+    if response.status_code > 200:
+        return response
+
+    # add this genomic_id to the sample drs object, if available
+    for sample in clinical_samples:
+        genomic_contents = {"drs_uri": [f"{HOSTNAME}/{genomic_sample['id']}"],
+                            "name": sample['sample_name_in_file'], "id": genomic_sample['id']}
+        obj = {
+            "id": f"{sample['sample_registration_id']}",
+            "contents": [genomic_contents],
+            "cohort": dataset,
+            "description": "sample",
+            "version": "v1"
+        }
+
+        response = requests.get(f"{url}/{'sample_registration_id'}", headers=headers)
+        if response.status_code == 200:
+            obj = response.json()
+            obj['contents'].append(genomic_contents)
+
+        response = requests.post(url, json=obj, headers=headers)
+        if response.status_code != 200:
+            return response
+
+    # master object:
+    obj = {
+        "contents": [
+            {
+                "drs_uri": [
+                    f"drs://{HOSTNAME}/{genomic_sample['id']}"
+                ],
+                "name": genomic_sample['id'],
+                "id": genomic_sample["type"]
+            },
+            {
+                "drs_uri": [
+                    f"drs://{HOSTNAME}/{genomic_sample['index']}"
+                ],
+                "name": genomic_sample['index'],
+                "id": "index"
+            }
+        ],
+        "id": genomic_sample['id'],
+        "name": genomic_sample['id'],
+        "description": "wgs",
+        "cohort": dataset,
+        "version": "v1"
+    }
+
+    for sample in clinical_samples:
+        clinical_obj = {"drs_uri": [f"{HOSTNAME}/{sample['sample_registration_id']}"],
+                        "name": f"{sample['sample_registration_id']}",
+                       "id": sample['sample_name_in_file'] }
+        obj["contents"].append(clinical_obj)
+
+    response = requests.post(url, json=obj, headers=headers)
+    return response
 
 
 def create_s3_sample(genomic_id: str, index: str, client):
